Employee Management System
Employee Information Management: This module provides a consolidated database for application tracking, employee demographics, compensation and benefit options, time tracking, and more. It’s also the central location for all of your personnel data, including the most up-to-date statistics on your organization’s hiring and retention patterns.
Employee Schedule Monitoring: Employee schedule monitoring entails the use of a variety of workplace surveillance techniques to obtain information on employees’ activities and whereabouts. Businesses keep tabs on employees in order to boost efficiency and safeguard company assets.
Monitor Working Days and Holidays: This is the practice of employers watching over their employees’ activities in the workplace and working days in order to ensure that they are productive. It will also help them to be aware when should their customer be active or take a break.